Update to v6.4.3.p4.6 whilst rebuilding the package
This commit aligns the build a bit more with the standard Debian way,
so the usual tools and CI can be confidently used.

This commit also removes a lot of the logic around the libvulkan
soname that isn't necessary anymore after changes done by the binary
provider.

From a coarse inspection it seems we're now aligned with how
meta-freescale organizes these libs.

Signed-off-by: Leonardo Held <leonardo.held@toradex.com>